tests/research_deepmind_weathernext.py
#!/usr/bin/env python3
"""
Research script for Google DeepMind's WeatherNext and GraphCast ensemble models.
These are state-of-the-art AI weather forecasting models.
"""

import requests
import json
import os
from datetime import datetime, timedelta

def research_deepmind_models():
    """Research available DeepMind weather models and access methods."""
    
    print("=" * 80)
    print("GOOGLE DEEPMIND WEATHER MODELS RESEARCH")
    print("=" * 80)
    
    print("\n1. AVAILABLE DEEPMIND WEATHER MODELS:")
    print("   ")
    print("   A. GraphCast (2023):")
    print("   - Paper: 'Learning skillful medium-range global weather forecasting'")
    print("   - GitHub: https://github.com/deepmind/graphcast")
    print("   - Model: Graph neural network for global weather forecasting")
    print("   - Resolution: 0.25° (25km)")
    print("   - Lead time: Up to 10 days")
    print("   - Variables: 227 atmospheric variables")
    print("   ")
    print("   B. WeatherNext (2024):")
    print("   - Latest ensemble forecasting model")
    print("   - Multiple ensemble members")
    print("   - Improved hurricane track prediction")
    print("   - Higher resolution than GraphCast")
    print("   ")
    print("   C. MetNet-3 (2023):")
    print("   - Regional weather forecasting")
    print("   - High-resolution precipitation")
    print("   - Real-time forecasting")
    
    print("\n2. ACCESS METHODS:")
    print("   ")
    print("   A. Open Source (GraphCast):")
    print("   - GitHub repository available")
    print("   - Can run locally with own data")
    print("   - Requires ERA5 input data")
    print("   - Python implementation")
    print("   ")
    print("   B. Google Cloud AI Platform:")
    print("   - WeatherNext API access")
    print("   - Requires Google Cloud account")
    print("   - Pay-per-use pricing")
    print("   - REST API interface")
    print("   ")
    print("   C. Research Collaborations:")
    print("   - Academic partnerships")
    print("   - Research access programs")
    print("   - Limited availability")
    
    print("\n3. GRAPHCAST OPEN SOURCE ACCESS:")
    print("   ")
    print("   Installation:")
    print("   ```bash")
    print("   pip install graphcast")
    print("   # or")
    print("   git clone https://github.com/deepmind/graphcast")
    print("   cd graphcast")
    print("   pip install -e .")
    print("   ```")
    print("   ")
    print("   Usage:")
    print("   ```python")
    print("   import graphcast")
    print("   from graphcast import GraphCast")
    print("   ")
    print("   # Load pre-trained model")
    print("   model = GraphCast.from_pretrained()")
    print("   ")
    print("   # Make predictions")
    print("   predictions = model.predict(inputs)")
    print("   ```")
    
    print("\n4. WEATHERNEXT API ACCESS:")
    print("   ")
    print("   Google Cloud Setup:")
    print("   1. Create Google Cloud project")
    print("   2. Enable WeatherNext API")
    print("   3. Set up authentication")
    print("   4. Get API credentials")
    print("   ")
    print("   API Endpoints:")
    print("   - Base URL: https://weathernext.googleapis.com/")
    print("   - Ensemble forecasts: /v1/ensemble")
    print("   - Single forecasts: /v1/forecast")
    print("   - Historical data: /v1/historical")
    
    print("\n5. HURRICANE FORECASTING CAPABILITIES:")
    print("   ")
    print("   GraphCast:")
    print("   - Global tropical cyclone tracking")
    print("   - 10-day forecasts")
    print("   - Multiple atmospheric variables")
    print("   - Requires ERA5 input data")
    print("   ")
    print("   WeatherNext:")
    print("   - Enhanced hurricane ensemble forecasts")
    print("   - Multiple ensemble members")
    print("   - Higher resolution for tropical storms")
    print("   - Real-time operational forecasts")
    
    print("\n6. DATA REQUIREMENTS:")
    print("   ")
    print("   Input Data:")
    print("   - ERA5 reanalysis data")
    print("   - Surface and pressure level variables")
    print("   - Global coverage")
    print("   - 6-hourly time steps")
    print("   ")
    print("   Output Data:")
    print("   - 227 atmospheric variables")
    print("   - 0.25° resolution")
    print("   - 10-day forecasts")
    print("   - Multiple lead times")
    
    print("\n7. COMPARISON WITH OTHER MODELS:")
    print("   ")
    print("   | Model        | Type      | Resolution | Lead Time | Access    |")
    print("   |--------------|-----------|------------|-----------|-----------|")
    print("   | ECMWF ENS    | Physics   | 18km       | 15 days   | MARS      |")
    print("   | NOAA GEFS    | Physics   | 25km       | 16 days   | Free      |")
    print("   | GraphCast    | AI        | 25km       | 10 days   | Open      |")
    print("   | WeatherNext  | AI        | 10km       | 10 days   | Cloud     |")
    
    print("\n8. RECOMMENDATIONS:")
    print("   ")
    print("   For Hurricane Eta/Iota (2020):")
    print("   1. GraphCast (Open Source)")
    print("      - Free, can run locally")
    print("      - Requires ERA5 input data")
    print("      - Good for methodology development")
    print("   ")
    print("   2. WeatherNext (Cloud API)")
    print("      - Latest technology")
    print("      - May have 2020 data")
    print("      - Requires Google Cloud setup")
    print("   ")
    print("   3. Hybrid Approach")
    print("      - Use GraphCast for methodology")
    print("      - Compare with WeatherNext if available")
    print("      - Validate against IBTrACS tracks")

def test_graphcast_installation():
    """Test if GraphCast can be installed and accessed."""
    
    print("\n" + "=" * 50)
    print("TESTING GRAPHCAST INSTALLATION")
    print("=" * 50)
    
    try:
        # Try to import graphcast
        print("Testing GraphCast import...")
        import graphcast
        print("✓ GraphCast successfully imported")
        
        # Check version
        print(f"GraphCast version: {graphcast.__version__}")
        
        return True
        
    except ImportError as e:
        print(f"✗ GraphCast not installed: {e}")
        print("\nTo install GraphCast:")
        print("pip install graphcast")
        print("or")
        print("git clone https://github.com/deepmind/graphcast")
        print("cd graphcast")
        print("pip install -e .")
        return False
    except Exception as e:
        print(f"✗ Error importing GraphCast: {e}")
        return False

def test_weathernext_api():
    """Test WeatherNext API access (if credentials available)."""
    
    print("\n" + "=" * 50)
    print("TESTING WEATHERNEXT API ACCESS")
    print("=" * 50)
    
    # Check for Google Cloud credentials
    google_creds = os.environ.get('GOOGLE_APPLICATION_CREDENTIALS')
    if google_creds:
        print(f"✓ Google Cloud credentials found: {google_creds}")
        
        try:
            from google.cloud import aiplatform
            print("✓ Google Cloud AI Platform imported")
            
            # Test API endpoint (this would require actual API access)
            print("Note: WeatherNext API access requires:")
            print("1. Google Cloud project setup")
            print("2. WeatherNext API enabled")
            print("3. Proper authentication")
            print("4. API quota/access permissions")
            
            return True
            
        except ImportError:
            print("✗ Google Cloud AI Platform not installed")
            print("Install with: pip install google-cloud-aiplatform")
            return False
    else:
        print("✗ No Google Cloud credentials found")
        print("Set GOOGLE_APPLICATION_CREDENTIALS environment variable")
        print("or run: gcloud auth application-default login")
        return False

def create_graphcast_example():
    """Create an example script for using GraphCast."""
    
    print("\n" + "=" * 50)
    print("GRAPHCAST EXAMPLE SCRIPT")
    print("=" * 50)
    
    example_code = '''#!/usr/bin/env python3
"""
Example script for using GraphCast for hurricane forecasting.
"""

import xarray as xr
import numpy as np
from datetime import datetime, timedelta

def setup_graphcast():
    """Set up GraphCast model for hurricane forecasting."""
    
    try:
        import graphcast
        from graphcast import GraphCast
        
        print("Loading GraphCast model...")
        model = GraphCast.from_pretrained()
        print("✓ GraphCast model loaded successfully")
        
        return model
        
    except Exception as e:
        print(f"✗ Error loading GraphCast: {e}")
        return None

def prepare_era5_input(hurricane_date="2020-11-01"):
    """Prepare ERA5 input data for GraphCast."""
    
    print(f"Preparing ERA5 input data for {hurricane_date}...")
    
    # This would require ERA5 data download
    # For hurricane analysis, we need:
    # - Surface variables (pressure, temperature, wind)
    # - Pressure level variables
    # - Global coverage
    # - 6-hourly time steps
    
    print("ERA5 data requirements:")
    print("- Surface variables: msl, 2t, 10u, 10v")
    print("- Pressure levels: z, t, q, u, v")
    print("- Coverage: Global")
    print("- Resolution: 0.25°")
    print("- Time: 6-hourly")
    
    return None

def run_graphcast_forecast(model, inputs):
    """Run GraphCast forecast for hurricane prediction."""
    
    print("Running GraphCast forecast...")
    
    try:
        # Make predictions
        predictions = model.predict(inputs)
        
        print("✓ GraphCast forecast completed")
        print(f"Forecast shape: {predictions.shape}")
        
        return predictions
        
    except Exception as e:
        print(f"✗ Error running GraphCast forecast: {e}")
        return None

def analyze_hurricane_tracks(predictions, hurricane_name="Eta"):
    """Analyze hurricane tracks from GraphCast predictions."""
    
    print(f"Analyzing {hurricane_name} tracks from GraphCast...")
    
    # Extract relevant variables
    # - Mean sea level pressure (for storm center)
    # - Wind fields (for intensity)
    # - Multiple lead times (for track evolution)
    
    print("Track analysis steps:")
    print("1. Extract storm center from pressure minima")
    print("2. Calculate wind speed from u/v components")
    print("3. Track storm position over time")
    print("4. Compare with IBTrACS observations")
    
    return None

if __name__ == "__main__":
    # Set up model
    model = setup_graphcast()
    
    if model:
        # Prepare input data
        inputs = prepare_era5_input()
        
        if inputs is not None:
            # Run forecast
            predictions = run_graphcast_forecast(model, inputs)
            
            if predictions is not None:
                # Analyze tracks
                analyze_hurricane_tracks(predictions)
'''
    
    print("Example GraphCast script created:")
    print(example_code)
    
    # Save to file
    with open("examples/graphcast_hurricane_example.py", "w") as f:
        f.write(example_code)
    
    print("\nSaved to: examples/graphcast_hurricane_example.py")

if __name__ == "__main__":
    research_deepmind_models()
    test_graphcast_installation()
    test_weathernext_api()
    create_graphcast_example() 